+++ cli/src/commands/auth/commands/login.ts
@@ -1,4 +1,4 @@
-import { existsSync, mkdirSync, readFileSync } from 'node:fs';
+import { mkdirSync } from 'node:fs';
 import { writeFile } from 'node:fs/promises';
 import { Command, program } from 'commander';
 import yaml from 'js-yaml';
@@ -6,21 +6,15 @@ import open from 'open';
 import pc from 'picocolors';
 import { configDir, configFile } from '../../../core/config.js';
 import { BaseCommandOptions } from '../../../core/types/types.js';
-import { performDeviceAuth, startPollingForAccessToken } from '../utils.js';
+import { checkConfigFile, performDeviceAuth, startPollingForAccessToken } from '../utils.js';
 
 export default (opts: BaseCommandOptions) => {
   const loginCommand = new Command('login');
-  loginCommand.description('Login a user.');
+  loginCommand.description('Login as an user.');
 
   loginCommand.action(async () => {
-    if (existsSync(configFile)) {
-      const data = yaml.load(readFileSync(configFile, 'utf8'));
-      const loginData = JSON.parse(JSON.stringify(data));
-      if (loginData && loginData?.expiresAt && new Date(loginData.expiresAt) > new Date()) {
-        console.log(pc.green('You are already logged in.'));
-        process.exit(0);
-      }
-    }
+    // checks if the config file exists and checks if the access token is not expired
+    checkConfigFile();
 
     const resp = await performDeviceAuth();
     if (!resp.success) {
